pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rembg==2.0.63",
    "flet==0.27.2",
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "pillow==11.1.0",
    "zenity==2.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]
